backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
numpy
opencv-python-headless
scikit-learn
reportlab
torch
torchvision
openai
faiss-cpu
sentence-transformers
python-magic
aiofiles
python-dotenv
Tesseract
pytesseract
pdfminer.six
efficientnet-pytorch
sqlalchemy
psycopg2-binary
alembic
python-jose[cryptography]
passlib[bcrypt]
python-multipart